# tp6/openapi/resas-api.yaml
openapi: 3.1.0
info:
  description: "Description de l'API du serveur de gestion des ressources de l'application \"MIF-Resas\".<br>Vous pouvez changer l'URL dans <code>servers</code> pour l'adapter au déploiement que vous souhaitez tester. Les autres valeurs ne doivent pas être modifiées."
  version: "1.0.0"
  title: "Mif-Resas"
...
servers:
  - url: "http://localhost:8080"
  - url: "http://localhost:8080/resas"
  - url: "http://192.168.75.xxx:8080/v3"
  - url: "https://192.168.75.xxx:8080/api/v3"
...
paths:
  /users:
    post:
      tags:
        - "Users"
      summary: "Créer un utilisateur"
      description: "Crée un utilisateur avec un login et nom. Remarque : il est créé si le login n'existait pas déjà. Dans ce cas, on renvoie un code 201. Sinon, on renvoie un 409 (Conflict)."
      operationId: "createUser"
      requestBody:
        description: "représentation de l'utilisateur"
        required: true
        content:
          'application/json':
            schema:
              $ref: "#/components/schemas/UserRequest"
            examples:
              application/json:
                value:
                  login: "toto"
                  name: "Toto"
          'application/xml':
            schema:
              $ref: "#/components/schemas/UserRequest"
          'application/x-www-form-urlencoded':
            schema:
              $ref: "#/components/schemas/UserRequest"
      responses:
        '201':
          description: "Utilisateur créé."
          headers:
            Location:
              schema:
                type: string
              description: "l'URL (locale) de l'utilisateur créé"
        '400':
          description: "Paramètres de la requête non acceptables"
        '409':
          description: "Un utilisateur avec ce login existe déjà"
    get:
      tags:
        - "Users"
      summary: "Récupérer la collection d'utilisateurs"
      description: "Renvoie les URI de tous les utilisateurs"
      operationId: "getAllUsers"
      responses:
        '200':
          description: "Liste des IDs de tous  les utilisateurs"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Users"
              example:
                users:
                  - "users/toto"
                  - "users/titi"
            application/xml:
              schema:
                $ref: "#/components/schemas/Users"
              example: "<users>\n\t<link href=\"users/toto\"/>\n\t<link href=\"users/titi\"/>\n</users>"
...
  /users/login:
    post:
      tags:
        - "Users"
      summary: "Connecter un utilisateur (opération métier)"
      description: "Connecte un utilisateur. Renvoie un code 204 si le login et le mot de passe (aka nom) correspondent, 401 sinon."
      operationId: "loginUser"
      requestBody:
        description: "login de l'utilisateur"
        required: true
        content:
          'application/json':
            schema:
              $ref: "#/components/schemas/UserRequest"
          'application/xml':
            schema:
              $ref: "#/components/schemas/UserRequest"
          'application/x-www-form-urlencoded':
            schema:
              $ref: "#/components/schemas/UserRequest"
      responses:
        '204':
          description: "Utilisateur logué."
          headers:
            Authorization:
              schema:
                type: string
              description: "le token d'authentification"
        '400':
          description: "Paramètres de la requête non acceptables"
        '401':
          description: "Les login et mot de passe ne correspondent pas"
        '404':
          description: "Utilisateur non trouvé"
  /users/logout:
    post:
      tags:
        - "Users"
      summary: "Déconnecter l'utilisateur connecté (opération métier)"
      description: "Déconnecte l'utilisateur sans l'enlever de la collection d'utilisateurs. Si vous passez l'authentification dans des headers HTTP, cette méthode ne fait rien mais répond que l'opération s'est bien passée."
      operationId: "logoutUser"
      responses:
        204:
          description: "Utilisateur déconnecté"
      security:
        - bearerAuth: []
...
components:
  schemas:
    Users:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/Link"
      xml:
        name: "users"
    UserRequest:
      type: object
      properties:
        login:
          type: string
          xml:
            name: "login"
        name:
          type: string
          xml:
            name: "name"
      xml:
        name: "user"
      required:
        - login
        - name
...
    Link:
      type: object
      properties:
        href:
          type: string
          description: The URL of the resource.
          xml:
            name: "href"
            attribute: true
      xml:
        name: "link"
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT